backend/requirements.txt
Django>=4.0
djangorestframework
hvac
cryptography
pycryptodome
oqs
python-decouple